[SECTION HEADER] Extract Assignment Header into reusable section header component
